AT Client change only: fix timeout wrap in uAtClientWaitCharacter().
The AT Client private function uATClientWaitCharacter() did not correctly handle ticks that happen after a wrap around.  This could lead to two failure cases:

- if the starting tick were in the range of (INT32_MAX - timeoutMs, INT32_MAX) the check for wrap around would causes the timeout to occur immediately,
- if the starting tick were in the range (INT32_MIN, 0) the timeout would only occur when the tick passes timeoutMs again, which could be as long as -INT32_MIN + timeoutMs ticks, or about 25 days.

If the module operates correctly, this should not strictly cause a problem.  However, a failure during the interval between INT32_MIN and zero, which does not provide sufficient characters, could take an unexpectedly long time to time out and fail.  There is also potential for the first case to cause a parsing failure in the AT Client which would result in a short response length being provided during a subsequent command attempt, triggering the second case.

Our thanks to warasilapm for spotting this and proposing the fix.